requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
msgpack==1.0.0
pycodestyle==2.6.0
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
